+++ addons/ui_react/editor_plugin/ui_react_dock.gd
@@ -520,6 +520,9 @@ func refresh() -> void:
 	else:
 		_issues_all.append_array(UiReactValidatorService.validate_nodes(nodes, root))
 
+	if root:
+		_issues_all.append_array(UiReactValidatorService.validate_wiring_under_root(root))
+
 	_issues_all.append_array(
 		UiReactUnusedStateService.build_issues(_dock_actions.resolve_output_dir(), root)
 	)